TA/bank_testing/pyproject.toml
[tool.poetry]
name = "bank_testing"
version = "0.1.0"
description = ""
authors = ["Butch78 <matthewtyleraylward@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
black = "^22.3.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
